refactor: split type mappers by entity into focused modules
Split _mapper_entities.py (200 lines) and mappers.py (193 lines) into
7 focused files, each under 100 lines. mappers.py is now a pure
re-export hub preserving the existing public API.
